Make diff answer a - b, signed and fractional
diff subtracted the wrong way round: a date two days ahead answered -2, the
opposite sign of every other library in this space, and the test that covered
it asserted only the magnitude with a comment saying the direction depended on
the engine — which is how it stayed unnoticed. It also returned an i64, so a
gap of five and a half months came back as five, the half silently gone.

Months and years are counted on the calendar rather than an averaged length:
whole steps first, then the remainder measured against the month it actually
falls in, so 15 days into February is 15/28 of a month and not 15/30. A month
reached by clamping counts as whole — January 31st plus one month is February
28th, so the gap between them is one month, where the truncating version said
zero.

Every value in the tests was taken from a reference implementation and matches
to the last digit. months_between had no caller left after the rewrite.

+++ crates/chronos-engine-napi/src/lib.rs
@@ -8,8 +8,8 @@ pub fn add(iso: String, amount: i64, unit: String) -> Result<String> {
 }
 
 #[napi]
-pub fn diff(a_iso: String, b_iso: String, unit: String) -> Result<i64> {
-  let result = catch_unwind(|| -> std::result::Result<i64, String> { chronos_engine::diff(&a_iso, &b_iso, &unit) });
+pub fn diff(a_iso: String, b_iso: String, unit: String) -> Result<f64> {
+  let result = catch_unwind(|| -> std::result::Result<f64, String> { chronos_engine::diff(&a_iso, &b_iso, &unit) });
   match result {
     Ok(Ok(value)) => Ok(value),
     Ok(Err(e)) => Err(Error::from_reason(e)),
@@ -106,7 +106,7 @@ pub fn end_of_in_zone(utc_iso: String, unit: String, zone: String) -> Result<Str
 }
 
 #[napi]
-pub fn diff_in_zone(a_utc: String, b_utc: String, unit: String, zone: String) -> Result<i64> {
+pub fn diff_in_zone(a_utc: String, b_utc: String, unit: String, zone: String) -> Result<f64> {
   let result = catch_unwind(|| chronos_engine::diff_in_zone(&a_utc, &b_utc, &unit, &zone));
   match result {
     Ok(Ok(v)) => Ok(v),
